pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdftwin"
version = "0.1.0"
description = "A developer-first tool to reconstruct any PDF into an editable IR and regenerate a visual twin."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Homer Quan" }
]
keywords = ["pdf", "cli", "ocr", "document-processing", "pymupdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pymupdf>=1.23.0",
    "litellm>=1.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "twine>=5.0.0",
]

[project.scripts]
pdftwin = "pdftwin.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pdftwin"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = true

[tool.pytest.ini_options]
addopts = "--cov=src/pdftwin --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
